move jitclasses out
