+++ ruff-lint-strict.toml
@@ -26,6 +26,7 @@ exclude = [
     "alphadia/data/alpharaw_wrapper.py",
     "alphadia/data/bruker.py",
     "alphadia/data/utils.py",
+    "alphadia/data/jitclasses/*.py",
     "alphadia/_fdrx/stats.py",
     "alphadia/_fdrx/plotting.py",
     "alphadia/_fdrx/base.py",
